move ugly code inside manager
