+++ antarest/study/storage/output_service.py
@@ -20,7 +20,6 @@ from antarest.core.exceptions import (
     OutputAlreadyArchived,
     OutputAlreadyUnarchived,
     OutputNotFound,
-    OutputVariablesViewError,
     TaskAlreadyRunning,
 )
 from antarest.core.filetransfer.model import FileDownloadTaskDTO
@@ -41,7 +40,7 @@ from antarest.study.business.output.utils import (
     MCIndAreasQueryFile,
     MCIndLinksQueryFile,
 )
-from antarest.study.business.output.variables_management import extract_variables_list
+from antarest.study.business.output.variables_management import check_variables_view_coherence, extract_variables_list
 from antarest.study.model import (
     ExportFormat,
     MatrixIndex,
@@ -69,108 +68,6 @@ from antarest.study.storage.utils import assert_permission, get_start_date, is_o
 from antarest.worker.archive_worker import ArchiveTaskArgs
 
 logger = logging.getLogger(__name__)
-
-
-def _checks_links_variables_view_coherence(
-    output_id: str, available_variables: OutputVariablesList, variable_name: str, area_from_id: str, area_to_id: str
-) -> None:
-    error_msg = f"The variable {variable_name} does not exist for link {area_from_id} - {area_to_id}."
-    link_variables = available_variables.mc_ind.links
-    for link_variable in link_variables:
-        if link_variable.area_1_name == area_from_id and link_variable.area_2_name == area_to_id:
-            if variable_name in link_variable.variables:
-                return
-            raise OutputVariablesViewError(output_id, error_msg)
-
-    raise OutputVariablesViewError(output_id, error_msg)
-
-
-def _checks_areas_variables_view_coherence(
-    output_id: str,
-    available_variables: OutputVariablesList,
-    variable_name: str,
-    variable_type: OutputVariablesType,
-    area_id: str,
-    thermal_id: str | None = None,
-    renewable_id: str | None = None,
-    st_storage_id: str | None = None,
-) -> None:
-    error_msg = f"The variable {variable_name} does not exist for area {area_id} and type {type}"
-    area_variables = available_variables.mc_ind.areas
-    for area_variable in area_variables:
-        if area_variable.name == area_id:
-            if variable_type == OutputVariablesType.THERMAL:
-                for thermal_variable in area_variable.thermal_clusters:
-                    if thermal_variable.name == thermal_id:
-                        if variable_name in thermal_variable.variables:
-                            return
-                        raise OutputVariablesViewError(output_id, error_msg)
-
-            elif variable_type == OutputVariablesType.RENEWABLE:
-                for renewable_variable in area_variable.renewable_clusters:
-                    if renewable_variable.name == renewable_id:
-                        if variable_name in renewable_variable.variables:
-                            return
-                        raise OutputVariablesViewError(output_id, error_msg)
-
-            elif variable_type == OutputVariablesType.SHORT_TERM_STORAGE:
-                for sts_variable in area_variable.short_term_storages:
-                    if sts_variable.name == st_storage_id:
-                        if variable_name in sts_variable.variables:
-                            return
-                        raise OutputVariablesViewError(output_id, error_msg)
-
-            else:
-                if variable_name in area_variable.variables:
-                    return
-                raise OutputVariablesViewError(output_id, error_msg)
-
-    raise OutputVariablesViewError(output_id, error_msg)
-
-
-def _checks_variables_view_coherence(
-    variable_type: OutputVariablesType,
-    output_id: str,
-    area_id: str | None = None,
-    area_from_id: str | None = None,
-    area_to_id: str | None = None,
-    thermal_id: str | None = None,
-    renewable_id: str | None = None,
-    st_storage_id: str | None = None,
-) -> None:
-    if variable_type == OutputVariablesType.LINK:
-        if any([area_id, thermal_id, renewable_id, st_storage_id]):
-            raise OutputVariablesViewError(output_id, "You provided an area related id for links")
-        if not area_from_id or not area_to_id:
-            raise OutputVariablesViewError(
-                output_id, "You should provide both `area_from_id` and `area_to_id` for links"
-            )
-        return
-
-    if any([area_from_id, area_to_id]):
-        raise OutputVariablesViewError(output_id, "You provided an link related id for areas")
-
-    if not area_id:
-        raise OutputVariablesViewError(output_id, "You should provide `area_id` for areas")
-
-    if variable_type == OutputVariablesType.THERMAL:
-        if not thermal_id:
-            raise OutputVariablesViewError(output_id, "You should provide `thermal_id` for thermal clusters")
-        if any([renewable_id, st_storage_id]):
-            raise OutputVariablesViewError(output_id, "You provided an storage/renewable id for thermal clusters")
-    elif variable_type == OutputVariablesType.RENEWABLE:
-        if not renewable_id:
-            raise OutputVariablesViewError(output_id, "You should provide `renewable_id` for renewable clusters")
-        if any([thermal_id, st_storage_id]):
-            raise OutputVariablesViewError(output_id, "You provided an storage/thermal id for renewable clusters")
-    elif variable_type == OutputVariablesType.SHORT_TERM_STORAGE:
-        if not st_storage_id:
-            raise OutputVariablesViewError(output_id, "You should provide `st_storage_id` for short-term storages")
-        if any([thermal_id, renewable_id]):
-            raise OutputVariablesViewError(output_id, "You provided an renewable/thermal id for short-term storages")
-    else:
-        if any([thermal_id, renewable_id, st_storage_id]):
-            raise OutputVariablesViewError(output_id, "You provided an renewable/thermal/storage id for areas")
 
 
 class OutputService:
@@ -733,29 +630,18 @@ class OutputService:
 
         # Checks the asked couple `variable name` / `object_id` exists for the output
         available_variables = self.get_output_variables_list(study_id, output_id)
-        _checks_variables_view_coherence(
-            variable_type, output_id, area_id, area_from_id, area_to_id, thermal_id, renewable_id, st_storage_id
+        check_variables_view_coherence(
+            output_id,
+            variable_type,
+            variable_name,
+            available_variables,
+            area_id,
+            area_from_id,
+            area_to_id,
+            thermal_id,
+            renewable_id,
+            st_storage_id,
         )
-
-        if variable_type == OutputVariablesType.LINK:
-            assert area_from_id is not None
-            assert area_to_id is not None
-            _checks_links_variables_view_coherence(
-                output_id, available_variables, variable_name, area_from_id, area_to_id
-            )
-
-        else:
-            assert area_id is not None
-            _checks_areas_variables_view_coherence(
-                output_id,
-                available_variables,
-                variable_name,
-                variable_type,
-                area_id,
-                thermal_id,
-                renewable_id,
-                st_storage_id,
-            )
 
         # Calls the aggregation with the right arguments
         # todo
